requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-phonenumber-field==6.3.0
phonenumbers==8.12.53
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.2
